delete key pairs when terminating

+++ bcli/bcterminate.py
@@ -11,9 +11,8 @@ def do_terminate(args):
     1. terminate all EC2 instances
     2. terminate all security groups
     """
-    file_deployed_info = bcutils.get_deployed_info_file()
-    if not os.path.exists(file_deployed_info):
-        print('deployed info file not found')
+    if bcutils.no_sessions():
+        print('no sessions, please deploy first')
         exit(1)
 
     nodes_info = bcutils.get_all_nodes_info()
@@ -22,6 +21,7 @@ def do_terminate(args):
         sgs_others = set()
         logging.info('Terminating resources in {} ...'.format(region_name))
         inst_id_list = [x['id'] for x in nodes_info[region_name]]
+        key_name = bcutils.get_key_pair_name()
         instances = awsutils.get_instances_info(region_name, inst_id_list)
         for item in instances:
             if item.state['Name'] == 'terminated':
@@ -39,6 +39,10 @@ def do_terminate(args):
             item.terminate()
             logging.info('- terminated {}'.format(item.id))
 
-        sgs_to_delete = list(sgs_to_delete)
-        awsutils.delete_security_groups(region_name, sgs_to_delete)
-        logging.info('- deleted security groups: {}'.format(sgs_to_delete))
+        awsutils.delete_key_pair(region_name, key_name)
+        logging.info('- deleted key pair {}'.format(key_name))
+
+        if sgs_to_delete:
+            sgs_to_delete = list(sgs_to_delete)
+            awsutils.delete_security_groups(region_name, sgs_to_delete)
+            logging.info('- deleted security groups: {}'.format(sgs_to_delete))
